Added files for the new approach
Now using the Darkflow-NNs, its output must be converted and the full data must be uploaded to labelbox.
Additionally the Frames shall be split into folders with 1000 images each. By now the images land in the same folder.

Alle scripts added, none is fully done. The Montage-Creation works but noch 100%